+++ sjf4j/src/test/java/org/sjf4j/schema/SchemaCoreTest.java
@@ -3,7 +3,6 @@ package org.sjf4j.schema;
 import lombok.extern.slf4j.Slf4j;
 import org.junit.jupiter.api.Test;
 import org.sjf4j.Sjf4j;
-import org.sjf4j.exception.SchemaException;
 
 import java.net.URI;
 import java.time.LocalDate;